+++ CoreOnly/Sources/Firebase.h
@@ -66,6 +66,10 @@ FirebaseAnalytics dependency to your project to ensure Firebase In App Messaging
     #endif
   #endif
 
+  #if __has_include(<FirebaseInstallations/FirebaseInstallations.h>)
+    #import <FirebaseInstallations/FirebaseInstallations.h>
+  #endif
+
   #if __has_include(<FirebaseInstanceID/FirebaseInstanceID.h>)
     #import <FirebaseInstanceID/FirebaseInstanceID.h>
   #endif