+++ FirebaseMessaging/Apps/Sample/Sample/SceneDelegate.swift
@@ -72,7 +72,7 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate, MessagingDelegate {
       .store(in: &cancellables)
   }
 
-  func messaging(_ messaging: Messaging, didReceiveRegistrationToken fcmToken: String) {
+  func messaging(_ messaging: Messaging, didReceiveRegistrationToken fcmToken: String?) {
     identity.token = fcmToken
   }
 }